+++ src/main/java/model/AttackStatus.java
@@ -1,0 +1,7 @@
+package model;
+
+public enum AttackStatus {
+	
+	WON, LOST, DRAW, FLAG_CAPTURE, ERROR;
+
+}