+++ src/main/java/model/GameStatus.java
@@ -1,0 +1,5 @@
+package model;
+
+public enum GameStatus {
+	MENU, RUNNING, PAUSED, WON;
+}